backend/requirements.txt
# Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
# psycopg2-binary==2.9.9  # Comentado temporariamente para usar SQLite
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# WebSockets & Real-time
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# File Upload & Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# Validation & Serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Monitoring & Debugging
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Utilities
python-decouple==3.8
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0

# Development
django-seed==0.3.1
factory-boy==3.3.0
faker==20.1.0

# Testing
pytest-django==4.7.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0 